+++ UnityTwitter/Assets/Twitter.cs
@@ -397,7 +397,7 @@ namespace Twitter
         {
             // Default implementation of UNIX time of the current UTC time
             TimeSpan ts = DateTime.UtcNow - new DateTime(1970, 1, 1, 0, 0, 0, 0);
-            return Convert.ToInt64(ts.TotalSeconds, CultureInfo.CurrentCulture).ToString(CultureInfo.CurrentCulture);
+            return Convert.ToInt64(ts.TotalSeconds, CultureInfo.InvariantCulture).ToString(CultureInfo.InvariantCulture);
         }
 
         private static string GenerateNonce()